backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
pydantic-settings==2.5.0
python-multipart==0.0.12
openpyxl==3.1.5
